Update New-WinSCPSession.ps1
There are some scenarios where the password is not required but just the UserName. A typical scenario regards the use of the privatekey. The Credential object is not easy to use because it cannot be created with a null or blank password.
You have to specify a dummy password using the following code:
$dummypassword = ConvertTo-SecureString "dummypwds"  -asplaintext -force
$credential = new-object -typename System.Management.Automation.PSCredential -argumentlist $user, $dummypassword
It is not user friendly, just to enter a username!
I propose to create a new parameter, UserName, to manage those scenarios instead of the Credential object.
Kind Regards