Add $XmlLogPath parameter to allow for specifying a custom location for the temporary WinSCP XML log file. (#155)
* Update New-WinSCPSession.ps1

Added $XmlLogPath parameter to allow for specifying a custom location for the XML Log Path. This is a workaround for issues I am encountering where the default Windows temp path cannot be written to.

* Update New-WinSCPSession.ps1

Removed redundant code setting the session XmlLogPath property in response to pull request feedback from tomohulk. The lines on 93-98 set all properties on $session as long as they have the same name as the function parameter name.

* Add $XmlLogPreserve parameter to New-WinSCPSession to allow setting of this session property in the function call.

Added $XmlLogPreserve to allow for specifying whether we want the XML log file to be retained or removed after process completion.

* Added valid path validation to the $XmlLogPath parameter.

Added validation to the new $XmlLogPath parameter using the method used to validate that it is a valid file path. Validation script is a copy of an above validation script.